zamieniono zmienne typu double na decimal

+++ AppLogic/ATMCard.cs
@@ -12,7 +12,7 @@ namespace AppLogic
         /// Tworzy nową kartę bankomatową o podanym numerze
         /// </summary>
         /// <param name="number">Numer karty</param>
-        public ATMCard(string number, Client owner, bool isActive, double balance) : base(number, owner, isActive, balance)
+        public ATMCard(string number, Client owner, bool isActive, decimal balance) : base(number, owner, isActive, balance)
         {
         }
 
@@ -20,14 +20,14 @@ namespace AppLogic
         {
         }
 
-        public override void MakeTransaction(double amount)
+        public override void MakeTransaction(decimal amount)
         {
             if (balance + amount < 0)
                 throw new InsufficientCardBalanceException("Próbowano pobrać większą kwotę niż to możliwe", this.Number, amount);
             balance += amount;
             Math.Round(balance, 2);
         }
-        public override BankActionResult Authorize(double amount)
+        public override BankActionResult Authorize(decimal amount)
         {
             if (balance + amount < 0)
                 return BankActionResult.REJECTED_INSUFFICIENT_ACCOUNT_BALANCE;
